Configurable custom bank interface.
